Fix cross compiling of docker libraries for beats
When trying to cross compile beats with the most recent version of docker the following platforms fail:

* solaris/amd64
* netbsd/amd64
* netbsd/386
* netbsd/arm

The reason is some tools missing for these platforms. Some of the code is used for running docker as a terminal client which is not used in Beats. The `func (p *JSONProgress) String` method now returns always an empty string. The method `CanAcess`was completely removed. Remove this content also removes quite a few dependencies which is great.

This is branched off `v17.03.1-ce`.